requirements.txt
PyQt5
pyi40aas
qtawesome